Persons

+++ server/controllers/personController.js
@@ -10,8 +10,8 @@ class PersonController {
     try {
       let { name, description, momId, dadId, userId } = req.body;
       const { image } = req.files;
-      // todo: transliterate file name, check ' '
-      let fileName = uuid.v4() + "-" + image.name;
+      // todo: transliterate file name
+      let fileName = uuid.v4() + "-" + image.name.replace(" ", "_");
       image.mv(path.resolve(__dirname, "..", "static", fileName));
       const person = await Person.create({
         name,
@@ -21,7 +21,6 @@ class PersonController {
         momId,
         dadId,
       });
-      console.log({ person });
       return res.json(person);
     } catch (e) {
       next(ApiError.badRequest(e.message));
@@ -30,26 +29,31 @@ class PersonController {
   async editOne(req, res, next) {
     try {
       const { name, description, momId, dadId, userId } = req.body;
-      console.log({ description });
       const { id } = req.params;
-      //console.log("reqFiles:", req.files);
-      //const { image } = req.files;
-      console.log({ id });
-      //console.log({ image });
-      console.log({ userId });
       // todo: transliterate file name, check ' '
-      const { image, createdAt } = await Person.findOne({ where: { id } });
-      console.log("personImage:", image);
-      console.log("reqFiles:", req.files);
+      const person = await Person.findOne({
+        where: { id },
+      });
+      if (+userId !== person.userId) {
+        next(
+          ApiError.forbidden(
+            "UserIds are not equal, you cannot edit this person"
+          )
+        );
+      }
+
       const newFileName =
-        req && req.files ? uuid.v4() + "-" + req.files.image.name : image;
+        req && req.files
+          ? uuid.v4() + "-" + req.files.image.name.replace(" ", "_")
+          : person.image;
 
       if (req && req.files) {
         await req.files.image.mv(
           path.resolve(__dirname, "..", "static", newFileName)
         );
-        await fs.unlink(path.resolve(__dirname, "..", "static", image), () =>
-          console.log(`file ${image} deleted`)
+        await fs.unlink(
+          path.resolve(__dirname, "..", "static", person.image),
+          () => console.log(`file ${person.image} deleted`)
         );
       }
 
@@ -62,7 +66,7 @@ class PersonController {
           momId,
           dadId,
         },
-        { where: { id } }
+        { where: { id, userId } }
       );
       //console.log({ updatedPerson });
       return res.json({
@@ -71,7 +75,7 @@ class PersonController {
         image: newFileName,
         momId,
         dadId,
-        createdAt,
+        createdAt: person.createdAt,
       });
     } catch (e) {
       next(ApiError.badRequest(e.message));
@@ -79,19 +83,17 @@ class PersonController {
   }
 
   async getAll(req, res) {
-    let { userId, limit, page } = req.query;
+    let { limit, page } = req.query;
     page = page || 1;
     limit = limit || 9;
     let offset = page * limit - limit;
     let persons;
-    if (userId) {
-      persons = await Person.findAndCountAll({
-        where: { userId },
-        limit,
-        offset,
-        order: [["id", "DESC"]],
-      });
-    }
+    persons = await Person.findAndCountAll({
+      limit,
+      offset,
+      order: [["createdAt", "DESC"]],
+    });
+
     //console.log({ persons });
 
     return res.json(persons);
@@ -103,20 +105,29 @@ class PersonController {
     //console.log("PERSON:", person);
     return res.json(person);
   }
-
   async deleteOne(req, res, next) {
     try {
       const { id } = req.params;
+      const { userId } = req.body;
+      console.log({ userId });
       const person = await Person.findOne({ where: { id } });
-      await person.destroy();
-      await fs.unlink(
-        path.resolve(__dirname, "..", "static", person.image),
-        () => console.log(`file ${person.image} deleted`)
-      );
-      return res.json({
-        id: person.id,
-        message: `Person ${person.name} was deleted from db`,
-      });
+      if (userId === person.userId) {
+        await person.destroy();
+        await fs.unlink(
+          path.resolve(__dirname, "..", "static", person.image),
+          () => console.log(`file ${person.image} deleted`)
+        );
+        return res.json({
+          id: person.id,
+          message: `Person ${person.name} was deleted from db`,
+        });
+      } else {
+        next(
+          ApiError.forbidden(
+            "UserIds are not equal, you cannot delete this person"
+          )
+        );
+      }
     } catch (e) {
       next(ApiError.badRequest(e.message));
     }
